requirements.txt
ffmpeg_python==0.2.0
numpy==1.24.4
openai_whisper==20240930
torch==2.4.1+cu118
